requirements-pi.txt
bitarray==1.7.1
PyYAML==5.1.1
scikit-build
redis==3.5.3
